feat: Support for custom remote names
This enables users to use craft to prepare releases targeting a remote
different than origin. For example, some people clone repositories such
that the remote 'origin' points to a fork and the remote 'upstream'
points to the upstream project repository.

Note that the string 'origin' is also found in ghPages.ts and
registry.ts, but those usages refer to a remote in a temporary clone
owned by craft and thus takes no customization.