feat(dry-run): Let dry run keep going after errors in `forEachChained` (#69)
The basic idea of dry-run mode is to go through the entire process, with no side effects, reporting what would happen and/or any errors which occur. Prior to this commit, however, errors inside of a `forEachChained` call would break the chain, so that the user wouldn't be able to see what would have happened had the rest of the links executed.

This commit fixes that by catching errors while executing each link. In order to accomplish the report-error-when-dry-run-throw-error-when-not behavior, it modifies the existing `reportError` function (which already does that with error _messages_) to also accept `Error` objects. Finally, it adds tests for each mode, with both sync and async iteratees.