fix(dry-run): Fix reporting of required artifact check in dry-run mode (#71)
In dry-run mode, errors don't actually stop iteration of a for loop (since they're just logged, not thrown), so it's possible to make it all the way through a fail-fast loop, even having failed on at least one iteration. Therefore, we can't assume that just because we made it past the for loop, the test succeeded.

This adds a flag and a check of that flag, so that we report the correct result in dry-run mode.