ref(zeus): Make zeus handle its own special cases (#72)
Prior to this, all of the `listArtifactsForRevision` methods on artifact providers (the base and its descendants) had a signature indicating that they returned one of three things: an array full of artifacts, an empty array, or `undefined`. The reason for this was historical, as these three possible return values corresponded to the three possible responses when asking zeus for a commit's artifacts, respectively

1. "Yup, got that commit, and here are the associated artifacts."
2. "Yup, got that commit, but there are no artifacts associated with it."
3. "Wait, what commit do you want stuff for? Never heard of it."

The only reason response number two was a possibility, though, was because zeus is both a status provider _and_ an artifact provider, the former role being the one which would let it learn of a commit independent of any associated artifacts; any service serving solely as an artifact provider would have no way of landing in situation two.

Further, zeus - unlike other artifact providers - will store multiple copies of the same file attached to the same commit, rather than overwriting the file when a newer version is uploaded to a commit which already has an older version. Therefore, when retrieving the file list, we need to group them by filename and sort them by update date, to mimic the overwriting behavior.

This change made it so that zeus, rather than the base class, has to deal with its own differing behavior, thereby simplifying the base class, its other descendants, and any files which use methods from same. Specifically, it let us cut `undefined` out of the signature for all of these methods, since we only need to represent situations one and three above, which we can do by returning either a full array or an empty array. (Zeus still handles all three cases, but now does so with log statements rather than return values, since functionally cases two and three are equivalent.)