fix(crates): Detect stale upload caches and retry (#98)
The publishing errors also happen locally. When publishing a crate with
`--no-verify`, it seems that the crate caches can go out of date and
`cargo` will fail to publish. We can detect this and retry publishing
with exponential backoff.

It is unclear what the root cause of this is. When cargo updates its
index, it runs a "fast path" request against the GitHub API which seems
to get cached. Waiting for 20-30 seconds seems sufficient to invalidate
those caches.