fix(github): Detect skipped status checks as successful (#124)
When GitHub Actions runs are skipped based on a condition, the status
check conclusion reports `skipped`. This is currently detected as a
failed status check by craft, where instead it should be considered
successful.

A use case of this is to skip certain status checks based on the kind of
a release. GitHub always reports all status checks and then marks them
as skipped.