fix(publish): Fix publishing when resuming from a state file
This is a follow up to #192. It fixes a bug that occurs when a
publish state file does not list all possible targets. In this case
any unlisted target, is skipped whereas its state should be treated
as "unpublished". This stemmed from a logic error in the if statement
which skipped publishing when the special target type "none" was passed.

The PR also refactors the code around this a bit to make it easier
to follow and less prone to errors. Ideal follow ups would be more
tests around publish and the state file functionality.