ref(logger): Move some extra info to debug level (#198)
Our default info level logging is a bit too noisy and outputs redundant information that would only be useful when debugging. This also caused issues withe the new `targets` command on CI as we were always outputting `[no-input]` when we detected CI. This PR moves most of these messages to debug level.