feat(config): Automatically detect GitHub config when missing
We don't need people to explicitly state `github` global configuration
as it can be detected from `git` itself. Since we rely on the presence
of a `git` repo in `publish` and `prepare` we can safely rely on this
too and remove another piece of redundant configuration.

This patch also adds tests around this.