build: Fix and simplify jest and TS configs
Our Jest and TS configs were a bit convoluted and since our TS config was based off of a web-based base config, it was building for browsers and was shipping/adding DOM-related things into our builds. This even uncovered a bug in our tests which is fixed with this patch.