ci: Fix our build matrix, add Node 14 & 16 (#211)
We were not setting the node version on CI build matrix which means we were just testing everything with Node 12 but twice. This patch fixes that, adds Node 14 & 16 to the build matrix, switches our artifact builds to be on the 'blessed' Node and Yarn versions defined in our package.json `volta` config, and bumps Yarn and Node to their latest versions without breaking changes.