fix(cli): Allow extra CRAFT_* env variables (#229)
Since we used 'strict mode' in `yargs` and then enabled mapping of _any_ `CRAFT_*` env variables to CLI args, Craft now complains when it sees env variables like `CRAFT_GCS_STORE_CREDS_JSON` which are only directly used from specific targets and not available through the CLI. In an ideal world, either these wouldn't start with the `CRAFT_` prefix, or would be available as CLI args. Since we are not in that world, we are doing a compromise: converting one simple thing we acn into a CLI arg, and then relaxing the strictness by only being strict about the commands but not options. Since this feature (`strictCommands()` was added in a later version, we also upgrade `yargs` to `15.4.1` which is the canonical version being used by other tools in the repo.

Fix up to #224.