fix(github-artifacts): Fix incorrect artifact resolution
Fixes #234. This is a follow up to #226. Although @chadwhitacre pointed out a logic issue in an earlier revision, we have failed to see that we were _always_ returning an artifact even if we don't have a proper name match as `reduce` without an initial value, starts with the first element of the array. This patch fixes the issue.