fix(logging): Fix scoped loggers not respecting log level (#236)
After the `consola` upgrade in #224, we stopped respecting the log level that is set further in execution in scoped logger which were created eagerly at module initialization time and not inheriting the log level that was changed afterwards. This patch moves these module-level loggers into base class constructors and standardizes them. This both fixes the issue and gives a common logging style for all artifact providers and targets.