feat(registry): Batch all registry targets
Fixes #48. This PR automatically batches all registry operations in a
concurrency-safe manner. It checks out a single, canonical copy of the
registry, adds all the commits there and pushes all of them at once
instead of doing piecemeal check-out and pushes.

The goals when preparing the patch were as follows:
- Do not require any existing config to change
- Do not require any other target to support batch-mode
- Do not add any RegistryTarget-specific code outside of that class
- Keep RegistryTarget class usable w/o a batch-helper or factory pattern