feat(publish): Ability to merge to non-default
This fixes #239 and fixes #115. This fix requires us to switch
from a purely GitHub API driven publish flow to a purely local
Git driven publish flow as it is not possible to compute the
nearest merge base via any GitHub API. We rely on `git log` and
some parsing to be able to determine this.