feat(changelog): GitHub Milestones based changelogs (#291)
This patch implements the GitHub Milestones based automatic
changelog generation. It builds up on #285 and #287 and ties
them with GitHub milestone data to generate a changelog that
is composed of logical sections with references to relevant
PRs.

Any remaining items are put under "Various fixes & improvements"
section as an itemized list.