feat(github): Verify release uploads for GitHub (#308)
We had an incident where we sent UTF-encoded string
when uploading GitHub release artifacts after an
Octokit upgrade due to misleading types. We haven't
noticed the issue until it was too late as we lack
basic checks such as file size and checksums. This
patch introduces both checks to the GitHub target.

Closes #303.