fix(crates): Resume workspace publish (#392)
The crates target is not able to resume a workspace publish because
already published crates cannot be republished. This means a partially
successful publish cannot be retried.

To work around this, the crates target can skip publishing a crate if it
already exists in this version. The downside of this is that we would
silence errors when publishing the wrong version.