fix(docker): Use proper Erlang binaries for our builds
The existing package paths for the compiled Erlang binaries
are no longer valid. This patch fixes them with the new ones
and fixes Docker image building.

Note that we are using the binaries compiled for Debian `bullseye`
instead of our `bookworm` version as they are not available yet.

One needs to check https://www.erlang-solutions.com/downloads/#