fix: Avoid global.json when running `dotnet push` (#601)
* fix: Avoid global.json when running `dotnet push`

Resolves #598:
- https://github.com/getsentry/craft/issues/598

Pinning the version of of the .NET SDK used to build the sentry-dotnet repo prevents us from making relases, since that version of .NET is not installed on the machine that publishes NuGet packages.

We can run `dotnet push` using any version of .NET so this PR simply runs that command from outside the repository working directory.

* Update nuget.ts